split up aliases files, check for file existence before sourcing separate bash files, add local project aliases to gitignore

+++ home/snippets.sh
@@ -1,11 +1,13 @@
 function snip {
   if [ -n "$1" ]; then
-    touch ~/Sites/castle/home/.sublime/"$1".sublime-snippet
-    cat ~/Sites/castle/home/.sublime/basic.sublime-snippet > ~/Sites/castle/home/.sublime/"$1".sublime-snippet
+    if [ ! -f ~/Sites/castle/home/.sublime/"$1".sublime-snippet ]; then
+      touch ~/Sites/castle/home/.sublime/"$1".sublime-snippet
+      cat ~/Sites/castle/home/.sublime/basic.sublime-snippet > ~/Sites/castle/home/.sublime/"$1".sublime-snippet
+    fi
     subl ~/Sites/castle/home/.sublime/"$1".sublime-snippet
   else
     echo "you need to put the name of the file you're trying to create"
-    echo "in the .sublime directory"
+    echo "in the snippets directory of .sublime"
     echo "path to: ~/Sites/castle/home/.sublime"
   fi
 }
